remove extension from pathing; move column naming
